+++ integration-tests/custom-allocator/lib.rs
@@ -114,7 +114,7 @@ mod custom_allocator {
         /// We test that we can upload and instantiate the contract using its default
         /// constructor.
         #[ink_e2e::test]
-        async fn default_works(mut client: ink_e2e::Client<C, E>) -> E2EResult<()> {
+        async fn default_works<Client: E2EBackend>(mut client: Client) -> E2EResult<()> {
             // Given
             let constructor = CustomAllocatorRef::default();
 
@@ -136,7 +136,7 @@ mod custom_allocator {
         /// We test that we can read and write a value from the on-chain contract
         /// contract.
         #[ink_e2e::test]
-        async fn it_works(mut client: ink_e2e::Client<C, E>) -> E2EResult<()> {
+        async fn it_works<Client: E2EBackend>(mut client: Client) -> E2EResult<()> {
             // Given
             let constructor = CustomAllocatorRef::new(false);
             let contract = client