+++ integration-tests/e2e-call-runtime/lib.rs
@@ -30,7 +30,9 @@ pub mod e2e_call_runtime {
         type E2EResult<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
         #[ink_e2e::test]
-        async fn call_runtime_works(mut client: ink_e2e::Client<C, E>) -> E2EResult<()> {
+        async fn call_runtime_works<Client: E2EBackend>(
+            mut client: Client,
+        ) -> E2EResult<()> {
             // given
             let constructor = ContractRef::new();
             let contract = client